Add minimum length for responses to be gzipped
Very short responses become larger if they are gzipped. With
the minLength parameter you can define what minimum length a
response should have before it is gzipped. The response is
buffered up to the minimum length. As soon as the length of
the response becomes larger than the minimum length, the buffered
part will be written to the gzip writer and the headers will be
set accordingly. After that no more buffering will be applied.
If the response stays shorter than the minimum length, it will
stay uncompressed and no Accept-Encoding and Vary headers will
be set.